fixed a possible bug involving size_t and loff_t
Previously the file offset was stored as an unsigned type (size_t).
However the offset should support negative values (loff_t).
For example SEEK_SET with -1 would have resulted in a very large
integer.  While this still would have been caught it was being
caught for being too large when it should have been because it was
too small.

+++ data_rw/data.c
@@ -15,7 +15,7 @@ struct device *data_device;
 struct data_dev {
 	struct cdev cdev;
 	char data[MAX_DATA];
-	size_t cur_ofs;  // current offset
+	loff_t cur_ofs;  // current offset
 } *data_devp;
 
 static int data_open(struct inode* inode, struct file* filp)
@@ -34,7 +34,7 @@ static ssize_t data_read(struct file *filp, char __user *buf, size_t count,
 					loff_t *f_pos)
 {
 	struct data_dev *data_devp = filp->private_data;
-	size_t cur_ofs;
+	loff_t cur_ofs;
 	char *datp;
 	size_t left;
 
@@ -57,7 +57,7 @@ static ssize_t data_write(struct file *filp, const char __user *buf, size_t coun
 					loff_t *f_pos)
 {
 	struct data_dev *data_devp = filp->private_data;
-	size_t cur_ofs;
+	loff_t cur_ofs;
 	char *datp;
 	size_t left;
